pack: Allow setting a custom image tag for Docker images
Previously, the image repository name was automatically computed from
the packages in the manifest without allowing the user to set a custom
one. As such, changing the packages in the manifest would result in a
new image name. Thereby requiring updating documentation et cetera when
using `docker load` directory on the resulting image.

Inspired by `docker build -t`, this commit adds a new Docker-specific
option to `guix pack` which allows setting a custom repository name for
the resulting image. If this option is not specified, pack falls back
to computing the name from the manifest. Therefore, this change is
entirely backwards compatible.

* guix/scripts/pack.scm (guix-pack): add --image-tag option
* guix/scripts/pack.scm (%docker-format-options): new constant
* guix/scripts/pack.scm (show-docker-format-options): new procedure
* guix/scripts/pack.scm (show-docker-format-options/detailed): new procedure
* guix/scripts/pack.scm (docker-image): allow setting a custom
  repository name for the created docker image via extra-options.
* doc/guix.texi (Invoking guix pack)[docker]: Document --image-tag option

Signed-off-by: Sören Tempel <soeren@soeren-tempel.net>