gnu: pixman: Disable non-optional verbose gnuplot debugging output
The --enable-gnuplot option was added previously in Guix Git commit
711316946fa84b55480681588f2cc23811c69784. The option causes any program
linked against the pixman library to constantly write gnuplot input to
standard output. I believe this is not desirable in the common case and,
in fact, this option is intended for testing purposes by upstream [1]. I
don't think that it should be enabled by default in a context of a
distribution.

I noticed this while working porting a Wayland status bar, which uses
pixman, to Guix and wondering why the status bar process was constantly
writing gnuplot input to standard output.

* gnu/packages/xdisorg.scm (pixman): Disable --enable-gnuplot.

[1]: https://gitlab.freedesktop.org/pixman/pixman/-/commit/d0e6c9f4f65e429058b97d2f947b048b445c17c4