ocaml: binsec: Wrap program to set OCAMLPATH
Otherwise, dune plugins cannot be loaded which results in some command
line options to not work correctly. For example, binsec --help errors
with "The library "binsec.sse.checkct" can't be found in the search
paths". To fix this, this patch adds a wrap-programs phase.

Fixes https://issues.guix.gnu.org/69996

* gnu/packages/ocaml.scm (binsec) <arguments>: Add wrap-programs phase.